+++ inputReader.hpp
@@ -13,7 +13,7 @@ struct InputEntry{
     int   fragmentId;
     int   x;
     int   y;
-    float rotation;
+    double rotation;
 
     std::string toString(void) const{
         return "[" + std::to_string(fragmentId) + "] (" + std::to_string(x) + "," + std::to_string(y) + ") (" + std::to_string(rotation) + ")";
